+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Permissions `view_same_corporation` and `view_same_alliance` now mean **all** corporations / alliances the user's characters belong to. Not only the main character.
 - Make badges fit into the menu
 
+### Fixed
+
+- Trying to delete a character from the admin site results in timeouts.
+
 ## [1.3.0] - 2021-04-17
 
 ### Added